% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{convert_human_to_mouse}
\alias{convert_human_to_mouse}
\title{Convert Human Gene Symbols to Mouse Equivalents via HomoloGene}
\usage{
convert_human_to_mouse(gene_list)
}
\arguments{
\item{gene_list}{Character vector of human gene symbols.}
}
\value{
Character vector of corresponding mouse gene symbols.
}
\description{
Convert Human Gene Symbols to Mouse Equivalents via HomoloGene
}
