% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{QRQ}
\alias{QRQ}
\title{Ordered Quantile Normalization followed by Z-Score Transformation}
\usage{
QRQ(Data)
}
\arguments{
\item{Data}{Numeric data frame or matrix.}
}
\value{
Normalized and standardized data frame.
}
\description{
Ordered Quantile Normalization followed by Z-Score Transformation
}
