% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFinx_BO.R
\name{SFinx_BO}
\alias{SFinx_BO}
\title{Spatial Fingerprint Binary Operator}
\usage{
SFinx_BO(
  Data,
  Finpre1,
  Finpre2,
  Thresh = 1,
  Pos,
  Image,
  Title = "Binary Analysis"
)
}
\arguments{
\item{Data}{Data frame containing both pathway scores.}

\item{Finpre1}{Column name of the first pathway fingerprint.}

\item{Finpre2}{Column name of the second pathway fingerprint.}

\item{Thresh}{Numeric threshold for |Z| activation.}

\item{Pos}{Data frame of coordinates (`x`, `y`).}

\item{Image}{Background tissue image.}

\item{Title}{Main title.}
}
\value{
List containing individual plots and combined grobs.
}
\description{
Spatial Fingerprint Binary Operator
}
