% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFinx_UO.R
\name{SFinx_UO}
\alias{SFinx_UO}
\title{Spatial Fingerprint Unary Operator}
\usage{
SFinx_UO(Data, Finpre, Thresh = 1, Pos, Image, Title = "SFinx Unary Analysis")
}
\arguments{
\item{Data}{Data frame containing the target pathway enrichment/expression scores.}

\item{Finpre}{Name of the target fingerprint column in `Data`.}

\item{Thresh}{Numeric threshold for |Z| score activation.}

\item{Pos}{Data frame containing spatial coordinates (`x` and `y`).}

\item{Image}{Magick/raster background image object.}

\item{Title}{Main plot title.}
}
\value{
A combined `ggplot` object displaying spot scores and 2D density maps.
}
\description{
Spatial Fingerprint Unary Operator
}
