% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ztran}
\alias{ztran}
\title{Z-score Standardization}
\usage{
ztran(x, na.rm = TRUE)
}
\arguments{
\item{x}{Matrix or data frame of numerical features.}

\item{na.rm}{Logical, whether to remove NA values.}
}
\value{
Z-transformed data frame/matrix.
}
\description{
Z-score Standardization
}
